fix(scheduler): run inject finalization in its own job
Inject dispatch crosses the network (RabbitMQ publish, external
executors) and a call that never returns parks the job thread. With
@DisallowConcurrentExecution Quartz then starts no further instance, so
the finalization sitting at the tail of execute() never runs again:
injects stay PENDING and simulations stay RUNNING indefinitely, with no
error surfaced anywhere.

Move handleInjectExpectationCollectStatus, handleAutoClosingSimulations
and handlePendingInject to InjectsFinalizationJob, triggered every
minute offset from the dispatch job. Their relative order is unchanged,
so the cross-tenant filter state each one relies on is preserved.

executeInTenant becomes the shared TenantScopedJobRunner, now used by
both jobs.
