+++ openaev-api/src/main/java/io/openaev/scheduler/jobs/InjectsExecutionJob.java
@@ -1,8 +1,6 @@
 package io.openaev.scheduler.jobs;
 
 import static io.openaev.aop.audit_log.AuditEventOrigin.SYSTEM;
-import static io.openaev.database.model.CollectExecutionStatus.COMPLETED;
-import static io.openaev.utils.inject_expectation_result.ExpectationResultBuilder.hasValidResults;
 import static java.time.Instant.now;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.groupingBy;
@@ -12,9 +10,6 @@ import io.openaev.aop.LogExecutionTime;
 import io.openaev.aop.audit_log.AuditEvent;
 import io.openaev.aop.audit_log.AuditEventScope;
 import io.openaev.aop.audit_log.AuditLogger;
-import io.openaev.context.TenantContext;
-import io.openaev.context.TenantScopedTransaction;
-import io.openaev.context.TxCtx;
 import io.openaev.database.model.*;
 import io.openaev.database.repository.ExerciseRepository;
 import io.openaev.database.repository.InjectDependenciesRepository;
@@ -26,19 +21,13 @@ import io.openaev.healthcheck.dto.HealthCheck;
 import io.openaev.healthcheck.utils.HealthCheckUtils;
 import io.openaev.helper.InjectHelper;
 import io.openaev.injector_contract.variables.contract.UserContract;
-import io.openaev.notification.model.NotificationEvent;
-import io.openaev.notification.model.NotificationEventType;
-import io.openaev.rest.exception.ElementNotFoundException;
 import io.openaev.rest.inject.service.AssetToExecute;
 import io.openaev.rest.inject.service.InjectService;
 import io.openaev.rest.inject.service.InjectStatusService;
+import io.openaev.scheduler.TenantScopedJobRunner;
 import io.openaev.scheduler.jobs.exception.ErrorMessagesPreExecutionException;
-import io.openaev.service.NotificationEventService;
-import io.openaev.service.SecurityCoverageSendJobService;
-import io.openaev.service.chaining.WorkflowService;
 import io.openaev.telemetry.metric_collectors.ActionMetricCollector;
 import io.openaev.utils.AgentUtils;
-import io.openaev.utils.ExecutionTraceUtils;
 import jakarta.persistence.EntityManager;
 import jakarta.validation.constraints.NotNull;
 import java.time.Instant;
@@ -60,7 +49,6 @@ import org.quartz.DisallowConcurrentExecution;
 import org.quartz.Job;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.expression.EvaluationContext;
 import org.springframework.expression.EvaluationException;
 import org.springframework.expression.Expression;
@@ -83,13 +71,6 @@ public class InjectsExecutionJob implements Job {
 
   private static final String ATOMIC_BATCH_KEY = "atomic";
 
-  @Value("${openaev.notification.simulation-completed-delay-seconds:3600}")
-  private long delayForSimulationCompletedEvent;
-
-  @Value(
-      "${inject.execution.threshold.minutes:" + DEFAULT_EXECUTION_THRESHOLD_TIME_IN_MINUTES + "}")
-  private Integer injectExecutionThreshold;
-
   private final InjectHelper injectHelper;
   private final InjectService injectService;
   private final ExerciseRepository exerciseRepository;
@@ -99,10 +80,8 @@ public class InjectsExecutionJob implements Job {
   private final InjectStatusService injectStatusService;
   private final io.openaev.executors.Executor executor;
   private final ActionMetricCollector actionMetricCollector;
-  private final NotificationEventService notificationEventService;
-  private final SecurityCoverageSendJobService securityCoverageSendJobService;
   private final EntityManager entityManager;
-  private final TenantScopedTransaction tenantTx;
+  private final TenantScopedJobRunner tenantScopedJobRunner;
 
   private final List<ExecutionStatus> executionStatusesNotReady =
       List.of(
@@ -114,7 +93,6 @@ public class InjectsExecutionJob implements Job {
   private final List<BaseInjectExpectation.EXPECTATION_STATUS> expectationStatusesSuccess =
       List.of(BaseInjectExpectation.EXPECTATION_STATUS.SUCCESS);
 
-  private final WorkflowService workflowService;
   private final HealthCheckUtils healthCheckUtils;
   private final Optional<AuditLogger> auditLogger;
 
@@ -134,110 +112,6 @@ public class InjectsExecutionJob implements Job {
         });
     exerciseRepository.saveAll(startedExercises);
     startedExercises.forEach(this::logScheduledLaunch);
-  }
-
-  public void handleAutoClosingSimulations() {
-    // Change status of finished simulations.
-    List<Exercise> mustBeFinishedSimulations = exerciseRepository.thatMustBeFinished();
-    // Filter out the simulations using the new chaining engine.
-    mustBeFinishedSimulations =
-        mustBeFinishedSimulations.stream()
-            .filter(simulation -> !workflowService.existsBySimulationId(simulation.getId()))
-            .toList();
-    if (mustBeFinishedSimulations.isEmpty()) {
-      return;
-    }
-
-    Map<String, List<String>> simulationIdsByTenant = new LinkedHashMap<>();
-    mustBeFinishedSimulations.forEach(
-        simulation ->
-            simulationIdsByTenant
-                .computeIfAbsent(simulation.getTenant().getId(), key -> new ArrayList<>())
-                .add(simulation.getId()));
-
-    simulationIdsByTenant.forEach(
-        (tenantId, simulationIds) ->
-            executeInTenant(
-                tenantId,
-                () -> {
-                  // Refetch in tenant scope so eager securityCoverage is resolved under multitenant
-                  // v2.
-                  List<Exercise> exercisesToFinish =
-                      new ArrayList<>(exerciseRepository.findAllById(simulationIds));
-                  exercisesToFinish.forEach(
-                      exercise -> {
-                        exercise.setStatus(ExerciseStatus.FINISHED);
-                        exercise.setEnd(now());
-                        exercise.setUpdatedAt(now());
-                      });
-                  List<Exercise> exercisesFinished = exerciseRepository.saveAll(exercisesToFinish);
-
-                  // maybe trigger stix coverage background job
-                  securityCoverageSendJobService.createOrUpdateCoverageSendJobForSimulationsIfReady(
-                      exercisesFinished);
-
-                  // send notification
-                  exercisesFinished.stream()
-                      .filter(
-                          ex ->
-                              ex.getScenario()
-                                  != null) // only send notification for exercise associated to a
-                      // scenario
-                      .forEach(
-                          ex ->
-                              notificationEventService.sendNotificationEventWithDelay(
-                                  NotificationEvent.builder()
-                                      .eventType(NotificationEventType.SIMULATION_COMPLETED)
-                                      .resourceType(ResourceType.SCENARIO)
-                                      .resourceId(ex.getScenario().getId())
-                                      .timestamp(Instant.now())
-                                      .build(),
-                                  delayForSimulationCompletedEvent));
-                }));
-  }
-
-  public void handlePendingInject() {
-    List<Inject> pendingInjects =
-        injectHelper.getAllPendingInjectsWithThresholdMinutes(this.injectExecutionThreshold);
-
-    if (pendingInjects.isEmpty()) {
-      return;
-    }
-
-    for (Inject inject : pendingInjects) {
-      InjectStatus status = inject.getStatus().orElseThrow(ElementNotFoundException::new);
-      // Find agents that already have a COMPLETE trace
-      Set<String> completedAgentIds = ExecutionTraceUtils.getCompletedAgentIds(status.getTraces());
-
-      // Get all agents expected to execute this inject
-      List<Agent> allAgents = injectService.getAgentsByInject(inject);
-
-      if (allAgents.isEmpty()) {
-        // Agentless inject: network scanners (e.g. Nuclei) target assets that have no agent, so the
-        // per-agent timeout loop below can never record anything. Without an explicit trace,
-        // updateFinalInjectStatus finalizes the inject ERROR from an empty COMPLETE-trace list and
-        // the execution details show only the initial "waiting to be consumed" info trace - a red
-        // inject with no reason. Add a clear agentless timeout trace instead, unless a terminal
-        // COMPLETE trace was already recorded (e.g. the injector reported the timeout itself).
-        boolean hasCompleteTrace =
-            status.getTraces().stream()
-                .anyMatch(t -> ExecutionTraceAction.COMPLETE.equals(t.getAction()));
-        if (!hasCompleteTrace) {
-          ExecutionTraceUtils.addAgentlessTimeoutTrace(status, this.injectExecutionThreshold);
-        }
-      } else {
-        // Add a COMPLETE/TIMEOUT trace for each agent that never responded
-        for (Agent agent : allAgents) {
-          if (!completedAgentIds.contains(agent.getId())) {
-            ExecutionTraceUtils.addTimeoutTrace(status, agent, this.injectExecutionThreshold);
-          }
-        }
-      }
-      injectStatusService.updateFinalInjectStatus(status);
-      // Save + stream one by one: the timeout finalization must reach the execution screens in
-      // real time (an inject stuck PENDING would otherwise stay "in flight" until a reload).
-      injectStatusService.saveAndStreamInject(status);
-    }
   }
 
   @VisibleForTesting
@@ -441,41 +315,6 @@ public class InjectsExecutionJob implements Job {
     exerciseRepository.save(exercise);
   }
 
-  /**
-   * Runs an inject execution under BOTH tenant scopes of the platform, set to the inject's tenant:
-   * the v2 primitive (transaction GUC, read by the inspector for activated tables such as
-   * collectors) and the v1 thread-local {@link TenantContext}, which {@link
-   * io.openaev.aop.HibernateFilterTransactionAspect} turns into the Hibernate {@code tenantFilter}
-   * on every {@code @Transactional} method it enters.
-   *
-   * <p>The v1 bridge is not optional: executing an inject resolves asset groups, endpoints and
-   * agents through Criteria queries, all still {@code @Filter} entities. {@link
-   * TenantContext#getCurrentTenant()} falls back to the DEFAULT tenant when the thread-local is
-   * unset, so without this a customer's simulation resolved the default tenant's endpoints and
-   * created its expectations against them - cross-tenant rows, and none for the real targets. It
-   * stayed invisible in single-tenant deployments, where that fallback happens to be the right
-   * tenant. Every other background executor (ScenarioExecutionJob, AtomicTestingExecutionJob,
-   * ExpectationsExpirationManagerJob, StepEventService...) already carries the same bridge.
-   *
-   * <p>Unlike those, this job runs on the shared {@code ForkJoinPool.commonPool} (nested {@code
-   * parallelStream}), which also borrows the calling thread: restore the previous value instead of
-   * clearing, so the scope of whatever else runs on that thread survives.
-   */
-  private void executeInTenant(@NotNull final String tenantId, @NotNull final Runnable work) {
-    String previousTenant =
-        TenantContext.hasCurrentTenant() ? TenantContext.getCurrentTenant() : null;
-    TenantContext.setCurrentTenant(tenantId);
-    try {
-      tenantTx.execute(TxCtx.forTenant(tenantId), work);
-    } finally {
-      if (previousTenant == null) {
-        TenantContext.clearCurrentTenant();
-      } else {
-        TenantContext.setCurrentTenant(previousTenant);
-      }
-    }
-  }
-
   @Override
   @LogExecutionTime
   public void execute(JobExecutionContext jobExecutionContext) throws JobExecutionException {
@@ -545,7 +384,7 @@ public class InjectsExecutionJob implements Job {
                     .forEach(
                         executableInject -> {
                           Inject inject = executableInject.getInjection().getInject();
-                          executeInTenant(
+                          tenantScopedJobRunner.runInTenant(
                               inject.getTenant().getId(),
                               () -> {
                                 try {
@@ -572,54 +411,14 @@ public class InjectsExecutionJob implements Job {
                               executableInject.getInjection().getInject().getTenant().getId())
                       .ifPresent(
                           tenantId ->
-                              executeInTenant(tenantId, () -> updateExercise(entry.getKey())));
+                              tenantScopedJobRunner.runInTenant(
+                                  tenantId, () -> updateExercise(entry.getKey())));
                 }
               });
-      // Change status of finished simulations.
-      handleInjectExpectationCollectStatus();
-      handleAutoClosingSimulations();
-      handlePendingInject();
     } catch (Exception e) {
       log.error(e.getMessage(), e);
       throw new JobExecutionException(e);
     }
-  }
-
-  @VisibleForTesting
-  void handleInjectExpectationCollectStatus() {
-    // Disable tenant filter — called from InjectsExecutionJob which runs cross-tenant
-    entityManager.unwrap(Session.class).disableFilter("tenantFilter");
-    List<Inject> injects = injectService.getExecutedAndNotFinished();
-    if (injects.isEmpty()) {
-      return;
-    }
-    List<Inject> fulfilled = new ArrayList<>();
-    for (Inject inject : injects) {
-      // An expectation is done collecting when it has nothing to collect (no result
-      // placeholders), when every result has been filled, or when its collection window has
-      // expired. The expiration escape is critical: partially filled expectations (one collector
-      // reported, another never did) keep empty placeholder rows forever and are not picked up by
-      // the expiration manager (their score is already set). Without it, a single silent
-      // collector leaves the inject COLLECTING and the simulation RUNNING indefinitely.
-      boolean collectDone =
-          inject.getExpectations().stream()
-              .allMatch(
-                  expectation -> {
-                    // Legacy expectation rows can carry a SQL NULL results column (see
-                    // InjectExpectationMapper): treat it as "nothing to collect" instead of
-                    // NPE-ing the job and blocking simulation auto-close.
-                    List<InjectExpectationResult> results = expectation.getResults();
-                    return results == null
-                        || results.isEmpty()
-                        || hasValidResults(results)
-                        || expectation.isExpired();
-                  });
-      if (collectDone) {
-        inject.setCollectExecutionStatus(COMPLETED);
-        fulfilled.add(inject);
-      }
-    }
-    injectService.saveAll(fulfilled);
   }
 
   // -- AUDIT LOGGING --
